add media-record.svg from gnome-icon-theme, as the basis for new icons
Taken from Debian git repository:
https://salsa.debian.org/gnome-team/gnome-icon-theme
commit 532e3019a19e16995e4e1087b49c4c38d4b4a7b6, Jul  5 20:22:40 2009 +0100

This particular file was never part of an official release: it was created
during the v2.28 development cycle but not included in it, and before the
next relase it was merged with all other media control icons into a single SVG.

This file is Copyrighted by Lapo Calamandrei, the same author of the Gnome
Disks SMART status icons currently in use by this project. The file was
originally licensed as GPLv2, but currently the whole gnome-icon-theme
project is dual-licenced GNU LGPL v.3 or Creative Commons BY-SA 3.0.